[inductor] Separate to_{dtype,device} from lowering to allow zero-copy
These lowerings must copy even when they are no-ops in order to preserve
correctness in the presense of mutations. However, `to_dtype` and `to_device`
are also used in various lowerings as a helper function where it is okay to alias.

So, I've split these into two functions and allow the helper functions to alias
which saves some unnecessary copies.

ghstack-source-id: a2fc1d82ca73872d40b8baca758365713620c48b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107640